(fix): move api_model classes inside package_api_model so that mlflow packaging works properly
